pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "openai-agents>=0.5.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
